+++ src/condor_schedd.V6/dedicated_scheduler.cpp
@@ -744,7 +744,7 @@ DedicatedScheddNegotiate::scheduler_handleJobRejected(PROC_ID job_id,char const 
 
 	SetAttributeInt(
 		job_id.cluster, job_id.proc,
-		ATTR_LAST_REJ_MATCH_TIME, (int)time(0), NONDURABLE);
+		ATTR_LAST_REJ_MATCH_TIME, time(0), NONDURABLE);
 }
 
 void
@@ -3826,7 +3826,7 @@ DedicatedScheduler::getUnusedTime( match_rec* mrec )
 			// This is the case we're really interested in.  We're
 			// claimed, but not active (a.k.a "Claimed/Idle").  We
 			// need to see how long we've been like this.
-		return( (int)time(0) - mrec->entered_current_status );
+		return( (int)(time(0) - mrec->entered_current_status) );
 		break;
 	default:
 		EXCEPT( "Unknown status in match rec %p (%d)", mrec, mrec->status );
